+++ conf/script/src/build_system/cmd/compiler/host/get_info/location/msvc/__init__.py
@@ -1,1 +1,1 @@
-from build_system.cmd.compiler.host.get_info.location.msvc.installation_path import find
+from build_system.cmd.compiler.host.get_info.location.msvc.impl import find
