+++ conf/script/src/build_system/cmd/compiler/host/get_info/location/msvc/impl.py
